Apply user checks on the specific endpoints in VUL-4.
This essentially reverts 447e3f928483f248d245aa556281fe621eb262b5, and
replaces it with a more specific implementation.

The most interesting part of this is probably the split in user types in
the frontend: the user area now only requests public fields, with the
other fields only being requested when the settings area becomes active.

Fixes #26063.