dev/sg: trim out buildkite log line prefix in stdout (#26069)
Turns out this issue is just the VS Code integrated terminal - logs appear to render fine in iTerm. However, since the cause of this issue in VS Code - the timestamps - are not really human-readable anyway, this PR just removes them anyway on '--out=stdout'. In the future we could introduce a flag to retain the timestamps if needed.

Co-authored-by: Chris Wendt <chrismwendt@gmail.com>