cmd: switch back to default repo for python3 (#26096)
It seems the alpine edge repo got updated introducing an incompatibility that is currently failing all builds on main. The original requirement for Python 3.9 was due to CVE-2021-29921, but switching back to the alpine 3.12 repo seems to pass the newly added Trivvy security scans which doesn't surface this CVE anymore, so we should be okay getting python3 from alpine 3.12 instead of alpine edge to fix these builds.