remove most uses of GlobalUsers
This removes most uses of the deprecated `database.GlobalUsers`,
replacing it with a db handle passed in from a parent scope. The
goal here is to inject the database dependency into all code that
uses it to facilitate mocking out database dependencies. The
remaining handful of references to `GlobalUsers` require slightly
more care to swap out, so I've left those for separate PRs so they
can be reviewed separately.