executorqueue: use google.golang.org/protobuf (#26281)
I was doing some "go mod" stuff and was getting warnings since
github.com/golang/protobuf is deprecated. This is a simple commit which
switched our only direct use of that module to the new one. Note:
timestamp.Timestamp is a type alias to timestamppb.Timestamp, so this is
actually the same thing.