codeintel: expose GraphQL API for API docs search
Builds on top of prior work such as #26973, which added the ability to actually
search the API docs search index at the store layer, by exposing a GraphQL API
which will be used to search API docs.

There are three reasons this is the right way to expose this API:

1. Debugging, having a direct interaction with this search backend that does not
   go through a bunch of other pipework is very useful.
2. Editor integration, which will invoke this API directly. In the future, I would
   like to have the editor analyze dependencies of the project and search just
   within those repositories by specifying them as parameters here. i.e. it is nice
   to be able to diverge from our standard search GraphQL API.
3. Experimentation, it's nice to have an isolated API we can tweak/adjust until
   things settle, without worrying about potential impacts to our real search API.

That said, it is likely I will also use the same codeintel resolvers here as the
backend for integrating these search API with our regular search suggestions API.

Testing: The logic here is small enough that I am not worried about testing this
right now. Obviously tests for this are a goal in the future, though.

Helps #21938

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>