lsifstore/apidocs: fix logic flaw in TextSearchQuery
The tests here clearly show it would join all query terms in sequence (AND) when in
reality we want it to join all query terms with a logical OR. This is because for a
query like:

```
public struct github.com/gorilla/mux mux.Router
```

Only the last term `mux.Router` matches the search key. All the other terms are
metadata terms and need to be ignored.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>