lsifstore/apidocs: reduce the amount of ':*' (match prefix) operators used
':*' (match prefix) operators [are VERY expensive.](https://github.com/sourcegraph/sourcegraph/pull/26487#issuecomment-949888182)

To reduce how many we use, we alter our emission of them so only the last lexeme
in a sequence gets a match-prefix operator. For example, if trying to match `github.com/sourcegraph/sourcegraph-testing`
then previously `git.com/source/source` would match it because it would effectively be
interpreted as `git*.*com*/*source*/*source*`. After this change, it would not, as
it will be interpreted as `git.com/source/source*`.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>