lsifstore: do not use dbconn.Global in tests
I came across this while putting together #26641 and realised that it's
pretty easy to switch to `dbtest` here.

Runtime difference: new version is ~300ms slower.

But I think it's worth it for now, since migrating with everything to
`dbtest.NewDB` brings us closer to the next step: optimizing
`dbtest.NewDB`.