internal/gitserver: Add method Client.httpPostWithURI
In this commit we introduce a set of internal API changes:

Client.do now accepts a URI instead of a path. It is now dumber (as it
should be) as a result and does not care if the path has an HTTP prefix
or not.

Client.httpPost will now apply the MD5 hashing scheme on the repo name
to determine the URI, before invoking Client.do.

A new method, Client.httpPostWithURI which does not apply any
transformations to the URI and sends the HTTP POST request as-is.