remove global TemporarySettingsStore mocks
This commit takes the least tied-up store (TemporarySettingsStore), and
replaces all uses of global mocks in tests with mocked calls on an
injected database.

The goal of this PR was to validate the approach, and ensure that I'm
not missing anything by transitioning one store completely. There are a
few items worth attention in the comments.