database: quiet external service migration errors
The OOB migrator added in #26759 to populate the `has_webhooks` field of
external services would error if the JSON configuration in the external
service record was invalid. It turns out that we have at least one
external service on cloud that has invalid JSON, as this wasn't always
enforced, so this means the OOB migration will never succeed.

Instead, we'll demote that to a log message and complete the migration
anyway.