search: remove support for names in configuration endpoint
Zoekt only sends repository IDs. This allows us to simplify the code and
make it purely based around repository IDs. I was tempted to update
types further here since it is a little clunky to be using int32 in the
backend code. However, there are some nicer refactors to do first before
I do that so that will come later.