search: extract excluded repos counting from repo resolution
This PR is a part of series to bring paginated repo resolution to
search.

It extracts the excluded repos counting step from repo resolution into
its own method, which is then called concurrently with other searches in
`doResults`.

This is needed because this counting doesn't make sense to do when
resolving repos in a paginated fashion.

Part of #26995