database: add Stars field to RepoName (#26994)
This PR is part of a series to migrate search to paginated repository
resolution.

This change is needed so that we can read the `Stars` value of the last
`types.RepoName` in a page of repos that is sorted by star count. That
value will then be put in a `database.Cursors` type which will be passed
around to a `searchrepos.Pager`.