database: remove unused PatternQuery
This was functionality I added a few years ago in pursuit of and/or
search queries. There was code which transformed a hierarchical search
query into WHERE clause against the repos table to help limit scope.

The idea of this package was to further extend it to create a more
expressive and less clunky alternative to the growing ReposListOptions
struct. However, the code has just remained unused so lets remove
it. Maybe someone will like this idea and pick it up again.