search: configuration handler part of indexer server
This is a more natural place for this handler to live. It will also be
more testable now which is the main motivation for these changes.

This commit is mostly a cut paste with less indentation. So the change
looks much bigger than it is. The only "functional" change is we go from
directly using "db" to using interfaces on the struct.