start pushing database.DB down from the top of the call tree
This puts database.DB at the very top of the call tree, from which we
can start propagating it downwards, meeting up with all the places I've
already started using it.

The goal here is to be able to inject a `database.DB` at the very top
level and pass that through the codebase from there. Right now, there
are still a number of references to `database.GlobalRepos` and
`dbconn.Global` which break this mockable chain, so in order to remove
the global db mocks completely, we'll need to get rid of those as well.
Unfortunately, getting rid of globals is difficult. Fortunately, I'm
feeling very motivated.