gitserver: Remove expensive count operation
This count needs to walk the entire table which isn't cheap on Cloud. Since the metric wasn't used in dashboards, I decided to just remove it. If you think this is of value, I suggest that we either use a fuzzy value here, or that we disable it on dotcom only.
Also, this metric was incorrect in the current state, because it didn't account for `options.OnlyWithoutShard = true` and used `blocked IS NULL`, which `IterateGitserverRepos` doesn't.

Q: Should we add the `blocked IS NULL` condition to `IterateGitserverRepos` as well?