search: introduce ConfigFingerprint
This is the opaque structure used by Sourcegraph to minimize the amount
of work done when zoekt polls for repository options. It is a timestamp
and a hash of the configuration. Using these two bits of information, we
can heuristically calculate what has changed since the fingerprint.

We version the fingerprint format since we expect changes in the future
as the requirements of inputs to IndexOptions changes.

Note: this only introduces the structure, but does not use it yet.