dbtesting: detect pkg name (#27073)
Using dbtesting is not how new code is written, so direct uses of this
should fade away over time. However, it still frustrates me to see its
tentacles all over our codebase via the init calls that set DBNameSuffix
global.

This is some fun code which eliminates the need to set DBNameSuffix by
detecting the package under test and generating the DBNameSuffix based
on that.

We also delete all uses of DBNameSuffix.