search: Repos.MinLastChanged takes into account search contexts
Zoekt polls Sourcegraph to find out which repositories to index. We
introduced MinLastChanged so that when polling it only needs to then
check gitserver for those repos which have changed since a certain
time. However, part of what could change an indexed repos is it being
included in a search context. As such we updated MinLastChanged to also
include any repository in a search context which has changed.

We confirm that updated_at is correctly updated anytime the set of
repositories changes.

Note: this is just a heuristic, zoekt will still occasionally poll
everything and not use MinLastChanged.