Centralize code for redaction and unredaction.
Creating interior willy-nilly pointers creates problems when trying to
use a type which has pointers of its own (JVMPackagesConnection).

The previous version of the test first takes an interior pointer for
the connection struct, then zeroes the struct, followed by an unmarshal
into the struct.

When pure structs are involved, this is fine, because the interior
pointer now points into the updated connection struct.

When a pointer field is involved, it gets set to nil, instead of the
struct being recursively zeroed, which means that the interior pointer
now points inside the old copy of the struct, instead of the updated
struct created during unmarshaling. Hence, any changes to fields cannot
be detected later.

With this change, we delay creating the interior pointer until the point
where we actually want to modify the struct.

As a bonus, combining the code paths for redaction and unredaction means
future changes can be focused in one place, instead of two.

No functional change intended.