search: paginated repository resolution (take 2) (#27359)
* search: implement paginated repository resolution

This PR is the last of series to introduce paginated repo resolution in
search.

We remove `backend.ListSearchable` and its use in repository resolution
because global searches on Sourcegraph.com no longer depend on it. This
in turn simplifies the code in preparation for cursor based pagination.

Part of #26995

* search: optimize RepoSearch job

This commit optimizes the RepoSearch job by leveraging the database to
find repos based on the text pattern. Yesterday's production incident
was caused by excessive database requests due to our previous brute-force
approach of loading all repos and then filtering them in the frontend.

* search: change repo stats semantics to repos with matches

This commit makes our repo stats semantics (in the UI and GraphQL API)
mean repositories with matches rather than searched or searchable.