update db schema for code monitor webhooks (#27474)
This commit updates the db schema to add tables for webhook
and slack notification for code monitors. It should not change any
behavior because the new tables will always be empty, the new
columns will always be nil, and the email column will still always be
non-nil. I will update the wrapping store in a followup PR.