replace uses of ReadActionEmailQuery with ListEmailActions
This replaces the remaining use of ReadActionEmailQuery with
ListEmailActions, and removes ReadActionEmailQuery. This also removes
one more instance of (CodeMonitorStore).Query().

This also gets rid of some unnecessary paging that was complicating the
changed function. We should be able to handle listing all email actions
configured for a monitor without needing to page through them.