authz: Remove sub-repo SupportedChecker (#27567)
Sub-repo permissions are additional filters applied after we have
already checked our existing repo permissions.

It is therefore not necessary to check whether sub-repo perms are
enabled for a repo. If they are, we'll have data in the
sub_repo_permissions table, which we plan to cache. If not, we assume
that the user has access to the whole repo based on our traditional repo
permissions checks.

SubRepoPermsClient was also exported as it is bad practice to have
constructors return private structs.

Co-authored-by: Robert Lin <robert@bobheadxi.dev>