unpaginate repos list for commit/diff search (#27834)
This collects the list of repos for commit/diff search before
executing the searches. This is necessary because we need to be able to
error if the number of repos to be searched is above the limit.
Currently, searches with too many repos will start searching, then time
out, returning no error. This is causing a massive increase in load on
gitserver.

This is a temporary measure, and will be reverted to the previous
implementation once we identify the bad code monitors that should be
disabled.