perf: Move publishing events to BigQuery and Amplitude to background
On dotcom, we would do these on the request path. Valery noticed that these requests disproportionally often fail with 504 (timeout) errors. After checking some logs, we found that some of these requests actually take up to a minute and more.
To our understanding, it is not critical that these happen on the request path, though. If they fail that would've been a problem before as well. Also, we now properly store away the local events all the time, which is a good improvement, I think.